+++ src/admin/templates/helpers/mod.rs
@@ -1,3 +1,4 @@
+mod admin_i18n;
 mod and;
 mod compare;
 mod concat;
@@ -17,6 +18,7 @@ use serde_json::Value;
 
 use crate::admin::Translations;
 
+use self::admin_i18n::AdminI18nHelper;
 use self::and::AndHelper;
 use self::compare::CompareHelper;
 use self::concat::ConcatHelper;
@@ -33,7 +35,13 @@ use self::translation::TranslationHelper;
 pub(super) fn register_helpers(hbs: &mut Handlebars, translations: Arc<Translations>) {
     hbs.register_helper("render_field", Box::new(RenderFieldHelper));
     hbs.register_helper("eq", Box::new(EqHelper));
-    hbs.register_helper("t", Box::new(TranslationHelper { translations }));
+    hbs.register_helper(
+        "t",
+        Box::new(TranslationHelper {
+            translations: translations.clone(),
+        }),
+    );
+    hbs.register_helper("admin_i18n", Box::new(AdminI18nHelper { translations }));
     hbs.register_helper("not", Box::new(NotHelper));
     hbs.register_helper("and", Box::new(AndHelper));
     hbs.register_helper("or", Box::new(OrHelper));
@@ -76,14 +84,18 @@ pub(super) fn as_f64(val: &Value) -> Option<f64> {
 
 #[cfg(test)]
 mod test_helpers {
-    use std::sync::Arc;
+    use std::{path::Path, sync::Arc};
 
     use crate::admin::{templates::create_handlebars, translations::Translations};
 
     pub fn test_hbs() -> handlebars::Handlebars<'static> {
         let tmp = tempfile::tempdir().expect("tempdir");
-        let translations = Arc::new(Translations::load(tmp.path()));
-        let hbs = create_handlebars(tmp.path(), false, translations).expect("create_handlebars");
+        test_hbs_with_translations(tmp.path())
+    }
+
+    pub fn test_hbs_with_translations(config_dir: &Path) -> handlebars::Handlebars<'static> {
+        let translations = Arc::new(Translations::load(config_dir));
+        let hbs = create_handlebars(config_dir, false, translations).expect("create_handlebars");
         (*hbs).clone()
     }
 }